refactor: centralize and reuse type definitions in types.ts for better consistency
